Add dryRun option; add proper logging in verbose 1 mode

+++ src/go/MONIT/ggus_alerting.go
@@ -169,6 +169,9 @@ func convertData(data ggus) []byte {
 		temp.StartsAt = _beginRFC3339
 		temp.EndsAt = time.Unix(maxtstmp, 0).UTC()
 
+		if verbose > 0 {
+			log.Println("adding", temp.Labels.Alertname, temp.Labels.Severity, temp.Labels.Service, temp.Labels.Tag, temp.Annotations.URL)
+		}
 		finalData = append(finalData, temp)
 
 	}
@@ -328,12 +331,16 @@ func categorizeTicket() {
 }
 
 //function containing all logics for alerting.
-func alert(inp string) {
+func alert(inp string, dryRun bool) {
 
 	jsonData := fetchJSON(inp)
 	var data ggus
 	data.parseJSON(jsonData)
 	jsonStrAM := convertData(data) //JSON data in AlertManager APIs format.
+	if dryRun {
+		fmt.Println(string(jsonStrAM))
+		return
+	}
 	post(jsonStrAM)
 	deleteAlerts()
 
@@ -345,11 +352,13 @@ func main() {
 	severity = "ticket" // TODO: replace with severity level from ticket annotation
 	tag = "monitoring"
 	service = "GGUS"
+	var dryRun bool
 
 	flag.StringVar(&inp, "input", "", "input filename")
 	flag.StringVar(&vo, "vo", "cms", "Required VO attribute in GGUS Ticket")
 	flag.StringVar(&alertManagerURL, "url", "", "alertmanager URL")
 	flag.IntVar(&verbose, "verbose", 0, "verbosity level")
+	flag.BoolVar(&dryRun, "dryRun", false, "dry run mode, fetch data but do not post it to AM")
 	flag.Parse()
 
 	if inp == "" {
@@ -366,5 +375,5 @@ func main() {
 		log.SetFlags(log.LstdFlags)
 	}
 
-	alert(inp)
+	alert(inp, dryRun)
 }
